requirements.txt
langchain>=0.2.21
langchain-community>=0.0.210
tiktoken>=0.4.0
sentence-transformers>=2.2.2
torch>=2.0.0
pypdf>=3.15.0
python-docx>=0.8.11
PyYAML>=6.0
pdf2docx>=0.5.8
chromadb>=0.4.4
ollama>=0.1.0
numpy<2.0.0
reportlab
requests>=2.25.0
beautifulsoup4>=4.9.0
urllib3>=1.26.0
sqlalchemy>=2.0.0
openpyxl>=3.1.5
docling==2.58.0
django>=5.0
djangorestframework>=3.15